refactor(portal): flatten Sidebar file structure
affects: @fremtind/portal
