+++ portal/src/components/Sidebar/Sidebar.scss
